% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{read.markdown}
\alias{read.markdown}
\alias{read_markdown}
\title{Read a markdown table into a data frame}
\usage{
read.markdown(file, stringsAsFactors = FALSE, strip.white = TRUE, ...)

read_markdown(file, trim_ws = TRUE, ...)
}
\arguments{
\item{file}{A path to a file, a connection, or literal data (either a single
string or a vector of lines).}

\item{stringsAsFactors}{Logical. Convert string columns to factors? Passed
along to \code{\link[=read.delim]{read.delim()}}.}

\item{strip.white, trim_ws}{Trim leading and trailing whitespace from each
column before parsing.}

\item{...}{Passed along to \code{\link[=read.delim]{read.delim()}} or \code{\link[readr:read_delim]{readr::read_delim()}} by
\code{read.markdown} and \code{read_markdown}, respectively. Applied after empty
and markdown-only lines have been removed.}
}
\value{
For \code{read.markdown}, a data.frame; for \code{read_markdown}, a tibble.
}
\description{
\code{read.markdown} and \code{read_markdown} are functions to read a markdown table
into a data.frame or tibble, respectively.
}
\details{
The \code{file} parameter will accept a filepath or connection, but given that
these functions are built for interactive use, they are built to accept a
single string containing the data (distinguished from a filepath by the
presence of a newline) or a vector of lines, as may be generated by
\code{\link[clipr:read_clip]{clipr::read_clip()}}
}
\examples{
read.markdown(
'+:--+--:+
 | x | y |
 +:==+==:+
 | 1 | 2 |
 | 3 | 4 |')

read_markdown(
'| x | y |
 | 1 | 2 |
 | 3 | 4 |')

}
