% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md.R
\name{read.md}
\alias{read.md}
\alias{read_md}
\title{Read a markdown table into a data frame}
\usage{
read.md(
  file = clipr::read_clip(),
  delim = "|",
  stringsAsFactors = FALSE,
  strip.white = TRUE,
  ...
)

read_md(file = clipr::read_clip(), delim = "|", trim_ws = TRUE, ...)
}
\arguments{
\item{file}{A path to a file, a connection, or literal data (either a single
string or a vector of lines). If unspecified, reads from the clipboard.}

\item{delim}{A string to use as a column delimiter. For most markdown
tables, this is the pipe: "|".}

\item{stringsAsFactors}{A logical value indicating whether to convert string
columns to factors. Passed along to \code{\link[=read.delim]{read.delim()}}.}

\item{strip.white, trim_ws}{Trim leading and trailing whitespace from each
column before parsing.}

\item{...}{Passed along to \code{\link[=read.delim]{read.delim()}} or \code{\link[readr:read_delim]{readr::read_delim()}} by
\code{read.md} and \code{read_md}, respectively. Applied after empty and
markdown-only lines have been removed.}
}
\value{
For \code{read.md}, a data.frame; for \code{read_md}, a tibble.
}
\description{
\code{read.md} and \code{read_md} are functions to read a markdown table into a
data.frame or tibble, respectively.
}
\details{
The \code{file} parameter will accept a filepath or connection, but given that
these functions are built for interactive use, they are built to accept a
single string containing the data (distinguished from a filepath by the
presence of a newline) or a vector of lines, as may be generated by
\code{\link[clipr:read_clip]{clipr::read_clip()}}.
}
\examples{
read.md(
'+:--+--:+
 | x | y |
 +:==+==:+
 | 1 | 2 |
 | 3 | 4 |')

read_md(
'| x | y |
 | 1 | 2 |
 | 3 | 4 |')

}
