% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{read_glimpse}
\alias{read_glimpse}
\alias{read.glimpse}
\title{Read the results of \code{tibble::glimpse} back into a data frame}
\usage{
read_glimpse(file = clipr::read_clip(), class = c("tbl_df", "tbl",
  "data.frame"))

read.glimpse(file = clipr::read_clip(), class = "data.frame")
}
\arguments{
\item{file}{Character. A path to a file, a connection, or literal
data (either a single string or a vector of lines). If unspecified,
reads from the clipboard.}

\item{class}{A character vector of classes to assign to the results.}
}
\value{
A data frame of the class specified by the \code{class} parameter.
}
\description{
\code{read_glimpse} and \code{read.glimpse} read the data printed
by \code{\link[tibble:glimpse]{tibble::glimpse()}} back into a data frame.
}
\details{
\code{read_glimpse} and \code{read.glimpse} are roughly inverses of \code{tibble::glimpse},
except:
\itemize{
\item they subset to rows where the input contains the complete data for the
observation,
\item list and matrix columns are unsupported and may lead to unexpected
behavior, and
\item since \code{glimpse} does not provide attributes or data frame classes,
\itemize{
\item attributes are dropped, and
\item class is assigned as supplied when called, defaulting to a tibble for
\code{read_glimpse} and a data.frame for \code{read.glimpse}.
}
}
}
\examples{

x <- capture.output(tibble::glimpse(iris))

read_glimpse(x)

read.glimpse(x)

}
