% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.so.R
\name{write.so}
\alias{write.so}
\alias{write_so}
\title{Write a text representation of a data frame}
\usage{
write.so(x, file = stdout(), write_clip = TRUE, indent = 4,
  tbl_fun = c("data_frame", "tibble"))

write_so(x, file = stdout(), write_clip = TRUE, indent = 4,
  tbl_fun = c("data_frame", "tibble"))
}
\arguments{
\item{x}{A data frame to write.}

\item{file}{A connection or filename to which to print; passed to \code{\link{cat}}.}

\item{write_clip}{Logical. Write result to clipboard?}

\item{indent}{The number of spaces by which to indent column definitions.}

\item{tbl_fun}{The character function name to use to create tibbles, either
\code{"data_frame"} or \code{"tibble"}. Ignored for other cases.}
}
\value{
Prints the call to produce the input to the specified connection;
returns the call invisibly.
}
\description{
\code{write.so} writes a call to produce the specified data frame to the console
or a specified connection. The function used to recreate the data frame in
the output is determined by the class of the input. \code{write_so} is an alias.
}
\examples{
write.so(head(iris))

}
